Fixed mouse_x and mouse_y underflow

+++ src/input.rs
@@ -70,13 +70,13 @@ pub fn read_chars() -> SmallVec<[u8; 4]> {
 }
 
 /// Gets the absolute mouse position on the X axis.
-pub fn mouse_x() -> u16 {
-    unsafe { dos_like_sys::mousex() as u16 }
+pub fn mouse_x() -> i32 {
+    unsafe { dos_like_sys::mousex() }
 }
 
 /// Gets the absolute mouse position on the Y axis.
-pub fn mouse_y() -> u16 {
-    unsafe { dos_like_sys::mousey() as u16 }
+pub fn mouse_y() -> i32 {
+    unsafe { dos_like_sys::mousey() }
 }
 
 /// Gets the mouse relative position
